feature: Add support for (process) parallelized tests to before_all for minitest.
When the test suite is run in parallel, the run method only enqueues the
test jobs and returns right after that, so we mustn't roll back the
transaction at that point.
Instead, intercept the Minitest.run_one_method call that actually runs
a test case. In each process, we will keep track of when the last test
case of a test file has run, and roll back the transaction in that moment.
This works because test cases from different test files are added to
the queue and processed in order, i.e. once we see a test case from a 
different class/test file, no test cases from the previous test file will
be run anymore.